services/auto_art_service/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
httpx==0.24.1
python-dotenv==1.0.0
redis==4.6.0
rq==1.12.0
pydantic==1.10.13
